Add log statements to empty lines
Sometimes you have a function that's failing and you don't know where.
It can be handy to place debug messages between statements to see how
far the code is getting.

This patch supports adding a log entry to and empty line.
It adds an incrementing number to easily know which entry you got to
without needing to care about line numbers